Refactoring of trade drop tables and code.
