--- classes/external/tradedrop_exporter.php
@@ -1,56 +1,0 @@
-<?php
-// This file is part of Moodle - http://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Moodle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-
-/**
- * Persistent exporter.
- *
- * @package    block_stash
- * @copyright  2017 Adrian Greeve - adriangreeve.com
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-
-namespace block_stash\external;
-defined('MOODLE_INTERNAL') || die();
-
-use lang_string;
-use moodle_url;
-use renderer_base;
-
-/**
- * Persistent exporter class.
- *
- * @package    block_stash
- * @copyright  2017 Adrian Greeve - adriangreeve.com
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-class tradedrop_exporter extends persistent_exporter {
-
-    protected static function define_class() {
-        return 'block_stash\\tradedrop';
-    }
-
-    protected static function define_related() {
-        return array('context' => 'context');
-    }
-
-    // protected static function define_other_properties() {
-    //     return [
-    //         'editurl' => [
-    //             'type' => PARAM_URL
-    //         ],
-    //     ];
-    // }
-}
