+++ db/upgrade.php
@@ -197,7 +197,7 @@ function xmldb_block_stash_upgrade($oldversion) {
         upgrade_block_savepoint(true, 2016082600, 'stash');
     }
 
-    if ($oldversion < 2017013001) {
+     if ($oldversion < 2017013004) {
 
         // Define table block_stash_trade to be created.
         $table = new xmldb_table('block_stash_trade');
@@ -206,6 +206,7 @@ function xmldb_block_stash_upgrade($oldversion) {
         $table->add_field('id', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, XMLDB_SEQUENCE, null);
         $table->add_field('stashid', XMLDB_TYPE_INTEGER, '11', null, XMLDB_NOTNULL, null, null);
         $table->add_field('name', XMLDB_TYPE_TEXT, null, null, null, null, null);
+        $table->add_field('hashcode', XMLDB_TYPE_CHAR, '40', null, XMLDB_NOTNULL, null, null);
         $table->add_field('timecreated', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
         $table->add_field('timemodified', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
         $table->add_field('usermodified', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
@@ -219,10 +220,10 @@ function xmldb_block_stash_upgrade($oldversion) {
         }
 
         // Stash savepoint reached.
-        upgrade_block_savepoint(true, 2017013001, 'stash');
+        upgrade_block_savepoint(true, 2017013004, 'stash');
     }
 
-    if ($oldversion < 2017013002) {
+    if ($oldversion < 2017013005) {
 
         // Define table block_stash_trade_items to be created.
         $table = new xmldb_table('block_stash_trade_items');
@@ -246,33 +247,7 @@ function xmldb_block_stash_upgrade($oldversion) {
         }
 
         // Stash savepoint reached.
-        upgrade_block_savepoint(true, 2017013002, 'stash');
-    }
-
-    if ($oldversion < 2017013003) {
-
-        // Define table block_stash_trade_drops to be created.
-        $table = new xmldb_table('block_stash_trade_drops');
-
-        // Adding fields to table block_stash_trade_drops.
-        $table->add_field('id', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, XMLDB_SEQUENCE, null);
-        $table->add_field('tradeid', XMLDB_TYPE_INTEGER, '11', null, XMLDB_NOTNULL, null, null);
-        $table->add_field('name', XMLDB_TYPE_TEXT, null, null, XMLDB_NOTNULL, null, null);
-        $table->add_field('hashcode', XMLDB_TYPE_CHAR, '40', null, XMLDB_NOTNULL, null, null);
-        $table->add_field('timecreated', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
-        $table->add_field('timemodified', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
-        $table->add_field('usermodified', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
-
-        // Adding keys to table block_stash_trade_drops.
-        $table->add_key('primary', XMLDB_KEY_PRIMARY, array('id'));
-
-        // Conditionally launch create table for block_stash_trade_drops.
-        if (!$dbman->table_exists($table)) {
-            $dbman->create_table($table);
-        }
-
-        // Stash savepoint reached.
-        upgrade_block_savepoint(true, 2017013003, 'stash');
+        upgrade_block_savepoint(true, 2017013005, 'stash');
     }
 
     return true;
